Add BooleanToVisibilityConverter class
Introduces the `BooleanToVisibilityConverter` class in the
`QuanLyThuVien.Resources.Converters` namespace. This class
implements the `IValueConverter` interface to convert boolean
values to `Visibility`, returning `Visible` for `true` and
`Collapsed` for `false`. The `ConvertBack` method is not
implemented, as reverse conversion is not supported. XML
documentation comments in Vietnamese are included to explain
the methods' functionality.